backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
redis>=5.0,<6.0
rq>=1.16,<2.0
langchain>=0.3,<0.4
langchain-openai>=0.2,<0.3
openai>=1.30,<2.0
pgvector>=0.2,<0.4
httpx>=0.27,<1.0
orjson>=3.10,<4.0
email-validator>=2.2,<3.0

